CameraServer/requirements.txt
Cython
Pillow
pyrealsense
numpy
matplotlib
mlxtend
dlib
opencv-python
